git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@505 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ ff_misc.h
@@ -60,7 +60,7 @@ void unlockHandle(FFHandle* hdl);
 // These functions are for code backwards compatibility:
 Handle PINEWHANDLE(int32 size);
 void PIDISPOSEHANDLE(Handle h);
-int32 PIGETHANDLESIZE(Handle h);
+size_t PIGETHANDLESIZE(Handle h);
 OSErr PISETHANDLESIZE(Handle h, int32 newSize);
 Ptr PILOCKHANDLE(Handle h, Boolean moveHigh);
 void PIUNLOCKHANDLE(Handle h);
